LPS-56224 - Fix of imports to class "BaseDDMTestCase" due the renaming of package "/test/unit/dynamicdatamapping".
