+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-test/test/unit/com/liferay/portlet/dynamic/data/mapping/io/DDMFormLayoutJSONDeserializerTest.java
@@ -15,7 +15,7 @@
 package com.liferay.portlet.dynamic.data.mapping.io;
 
 import com.liferay.portal.kernel.util.LocaleUtil;
-import com.liferay.portlet.dynamicdatamapping.BaseDDMTestCase;
+import com.liferay.portlet.dynamic.data.mapping.BaseDDMTestCase;
 import com.liferay.portlet.dynamicdatamapping.io.DDMFormLayoutJSONDeserializerImpl;
 import com.liferay.portlet.dynamicdatamapping.io.DDMFormLayoutJSONDeserializerUtil;
 import com.liferay.portlet.dynamicdatamapping.model.DDMFormLayout;
