+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-test/test/unit/com/liferay/portlet/dynamic/data/mapping/model/impl/DDMStructureImplTest.java
@@ -15,7 +15,7 @@
 package com.liferay.portlet.dynamic.data.mapping.model.impl;
 
 import com.liferay.portal.kernel.util.LocaleUtil;
-import com.liferay.portlet.dynamicdatamapping.BaseDDMTestCase;
+import com.liferay.portlet.dynamic.data.mapping.BaseDDMTestCase;
 import com.liferay.portlet.dynamicdatamapping.model.DDMForm;
 import com.liferay.portlet.dynamicdatamapping.model.DDMFormField;
 import com.liferay.portlet.dynamicdatamapping.model.DDMStructure;
