+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-test/test/unit/com/liferay/portlet/dynamic/data/mapping/storage/DDMFormFieldValueTest.java
@@ -15,7 +15,7 @@
 package com.liferay.portlet.dynamic.data.mapping.storage;
 
 import com.liferay.portal.kernel.util.StringUtil;
-import com.liferay.portlet.dynamicdatamapping.BaseDDMTestCase;
+import com.liferay.portlet.dynamic.data.mapping.BaseDDMTestCase;
 import com.liferay.portlet.dynamicdatamapping.model.UnlocalizedValue;
 import com.liferay.portlet.dynamicdatamapping.storage.DDMFormFieldValue;
